Simplify package set selection in anaconda
tab will both reach the environment selection list and then go
through it, so we don't need to start out with tab and then
switch to down. And since all we need to do is hit one key until
a needle matches, let's use the handy function os-autoinst
provides for doing this...

This should fix it on current Rawhide (where it only takes one
tab, or zero tabs - not sure which - to reach the list).

Signed-off-by: Adam Williamson <awilliam@redhat.com>