app start/stop tests: use normal flags for GNOME Terminal test
This test had special flags because it used to be run first and
did the prep steps for the other tests. Now the aaa_setup test
does that job, so there's no reason for this test to have these
flags any more, and somehow they seem to be breaking the test on
Rawhide. Let's give it the same flag as all the other normal
tests.

Signed-off-by: Adam Williamson <awilliam@redhat.com>