Tweak GNOME Software Updates needle match area slightly
It kinda looks like the text has been moved closer to the icon
in the latest gnome-software, such that the little '1' bubble
is now in the needle match area if it's present. Obviously that
may vary, so let's tweak the needle match area a bit to the
right so it's hopefully consistent and still matches on older
versions.