desktop_printing: wait a bit after switching back from console
This test is failing a lot recently on KDE because when we try
to launch konsole immediately after returning from a console to
the desktop, it is mistyped and we get a browser instead. Let's
try a little sleep after the switch back in case KDE just needs
to settle down a bit.

Signed-off-by: Adam Williamson <awilliam@redhat.com>